+++ resources.go
@@ -29,6 +29,43 @@ type resourceAPI interface {
 	GetFolders() ([]rainforest.Folder, error)
 	GetBrowsers() ([]rainforest.Browser, error)
 	GetSites() ([]rainforest.Site, error)
+	GetRunGroups() ([]rainforest.RunGroup, error)
+}
+
+func printRunGroupDetails(c cliContext) error {
+	runGroupIDStr := c.Args().First()
+	if runGroupIDStr == "" {
+		return cli.NewExitError("Please supply the Run Group ID. Use the `run-groups` command to retrieve a list of Run Group IDs.", 1)
+	}
+
+	runGroupID, err := strconv.Atoi(runGroupIDStr)
+	if err != nil {
+		return cli.NewExitError(err.Error(), 1)
+	}
+
+	runGroup, err := api.GetRunGroupDetails(runGroupID)
+	if err != nil {
+		return cli.NewExitError(err.Error(), 1)
+	}
+
+	runGroup.Print()
+	return nil
+}
+
+func printRunGroups(api resourceAPI) error {
+
+	runGroups, err := api.GetRunGroups()
+	if err != nil {
+		return cli.NewExitError(err.Error(), 1)
+	}
+
+	rows := make([][]string, len(runGroups))
+	for i, rungrp := range runGroups {
+		rows[i] = []string{strconv.Itoa(rungrp.ID), rungrp.Title}
+	}
+
+	printResourceTable([]string{"Run Group ID", "Run Group Name"}, rows)
+	return nil
 }
 
 // printFolders fetches and prints out the available folders from the API