+++ runner_test.go
@@ -101,6 +101,7 @@ func TestMakeRunParams(t *testing.T) {
 				"description":    "my awesome description",
 				"environment-id": "1337",
 				"tag":            []string{"tag", "tag2,tag3"},
+				"run-group-id":   "14",
 			},
 			args: cli.Args{"12", "34", "56, 78"},
 			expected: rainforest.RunParams{
@@ -113,6 +114,7 @@ func TestMakeRunParams(t *testing.T) {
 				EnvironmentID: 1337,
 				Tags:          []string{"tag", "tag2", "tag3"},
 				Tests:         []int{12, 34, 56, 78},
+				RunGroupID:    14,
 			},
 		},
 		{